qca9377-firmware-aml: change firmware version to 1.0.0-3
Firmware tar.gz available on linode always includes the latest version without changing the file name. This means that actual firmware version is 1.0.0-4, which is known to degrade performance, at least on Amlogic devices. To fix this regression and make sure that we always get the same firmware version, repack the 1.0.0-3 firmware and provide it from our server.