kodi: advancedsettings.xml disable Dirty Regions completely to fix Subtitles - video playback stuttering on S912's.
https://kodi.wiki/view/HOW-TO:Modify_dirty_regions
https://kodi.tv/article/working-dirty-regions

To be used with the r7p0 drivers commits.

Also CPU software decoding and rendering is much improved allowing for
example 1080p 11Mbps H264 software decoding to be used on S912's.
Which should be good for 1080p H264 Netflix and possbily Amazon Video.

A 792Mhz GPU clock may need to be used for 1080p H264