u-boot: update odroid-c2 to latest version
This required a slight refactoring of the build process, as hardkernel split the
firmware into two separate git repos.
The firmware part got its own package named u-boot_firmware which is now fetched as
dependency of the u-boot package when odroid-c2 is selected.
This was required in order to get the bl301 binary to compile with our patches.