kodi: enable games and disable rewind by default
We want to get more visibility for retroplayer so let's enable games by default.
The only reason kodi doesn't do this is because it doesn't have the sense of a
binary add-on repo. LibreELEC is special in that we have had this ability forever.

We also want to disable rewind by default because it makes low powered devices
suffer in performance. It also isn't exposed in a manner that makes it easy to use yet.