kodi/amlogic: silence error caused by race condition
Upon booting Kodi the following error appears, "Remote
communication server, failed to start" this is due to the network
not being fully ready before Kodi starts.

When the network is ready the server does start as normal anyway.

This can also be resolved by enabling "Wait for network before
starting Kodi" in CE-settings.