batocera-wifi: Fix parsing of SSIDs with spaces.
Instead of using regexp magic, this patch simply checks for an "wifi_"
string in the line, which is probably just needed when connmanctl
outputs error messages.

This patch only uses internal bash and does not rely on external
commands like `grep` and `sed`.

The SSID line is simply split and the last word, which is the connmanctl
identifier, is removed. So we end up with full SSID strings, even if
they contain spaces or other symbols the regexp did not predict.