scripts/get_archive: Allow the use of GitHub Tokens
You can define COREELEC_GHTOKEN in your ~/.coreelec/options file.
This is usefull if you are testing some development changes in a private
GitHub repo before pushing it to a public one.
This Token is sent in an additional Header only to https://github.com/.
Also this is restricted to packages that define PKG_USETOKEN=yes.

Co-authored-by: Raybuntu <57228613+Raybuntu@users.noreply.github.com>