init: fix label on flash partition on eMMC
fsck from newest dosfstools package removes the label on fat partition created with
some old version and system doesn't boot. The fix is checking the label after fsck
step and change it in case it was removed or it is wrong.
Also some output from init script and fsck is send to dmesg.

relevant output from dmesg
[    1.599622@3]- init: ### Checking disk(s):  LABEL=CE_FLASH /dev/CE_STORAGE
[    1.841261@0]- init: ### /dev/CE_STORAGE: clean, 14619/3514368 files, 1493991/14041600 blocks
[    1.841339@0]- init: ### fsck.fat 4.2 (2021-01-31)
[    1.841481@0]- init: ### Volume label 'CE_FLASH' stored in root directory is not valid.
[    1.841553@0]- init: ###   Auto-removing label.
[    1.841586@0]- init: ###
[    1.841672@0]- init: ### *** Filesystem was changed ***
[    1.841734@0]- init: ### Writing changes.
[    1.841850@0]- init: ### /dev/CE_FLASH: 20 files, 56461/130815 clusters
[    1.843001@0]- init: ### Filesystem errors corrected , continuing...
[    1.854708@0]- init: ### wrong eMMC flash label "", fixing it