Amlogic-old: Add alternative kernel option
USE AT YOUR OWN RISK NO SUPPORT WILL BE GIVEN IF YOU USE THIS KERNEL
PLEASE DO NOT OPEN ANY ISSUES ON GITHUB OR ASK FOR SUPPORT IN THE FORUM
or DISCORD. THIS KERNEL IS FROZEN NO MORE CHANGES WILL BE MADE IF IT 
WORKS FOR YOU IT WORKS

This is an alternative kernel for older Amlogic devices (s905, s905(?)) 
used specially in 1gb ram devices it can give a bit more performance on
certain emulators.

To use it just unzip this file on the root of the EMUELEC partition 
and replace the files.