options: raise default ccache limit to 20G (#1022)
The 10G ccache limit is enough for upstream projects (CE, LE), as
their build will need only ~5.9GiB of ccache. But we need ~11GiB
even for the current build. The 10G limit effectively renders our
ccache not usuable as we will continously chase the tail of ccache
due to its LRU (least recently used) strategy and our linear build
style (even with MT): When we need ccache for some packages, in
most of the cases it's already removed since it's considered old
for a latter pacakge that's built in the last build.

Raising the size limit to 20G should solve the ccache-chasing issue
for now, and will be future-proof for a long time, even if we add
more emulators/ports