linux: unify USB serial driver config
This was motivated by PL2303 being missing from RK3328 but I figured if
I am at it I might as well clean it up for all projects.
I took the config from Generic as it seemed sensible (includes all
common USB to serial dongles) and replicated it for all other projects.

For a few I disabled CONFIG_USB_SERIAL_GENERIC, that is only a testing
driver and should not be used productively anyways [1].

For Qualcomm I disabled a ton of drivers, but these were never
explicitly enabled in the first place, being active from the first
kernel config made for this project. Note that this change does not
affect platform drivers but only USB drivers so removing the Qualcomm
driver here is not a problem unless people want to debug other Qualcomm
hardware from their LibreELEC system.

Other than in those two instances the enabled options should be a
superset of what was previously enabled.

[1] https://elixir.bootlin.com/linux/v6.1.4/source/drivers/usb/serial/generic.c#L42