linux: update patches to fix audio passthrough issues
The ELD constraint patch is updated to a more correct version which
now properly sets the constraints based on rate families.

The added hdmi-codec patch fixes audio infoframe values for passthrough
(channel info needs to be set to "refer to stream header") as some
TVs (eg Sony) tripped over that and refused to play TrueHD otherwise.

Signed-off-by: Matthias Reichl <hias@horus.com>