mesa (RPi): temporarily drop B10G10R10A2 format support
This prevents kodi using 10bpc planes with only 2 bits alpha which
result in poor visual quality of (semi-) transparent GUI overlays
when using direct-to-plane rendering.

Ideally this should be fixed in kodi but the changes are likely quite
involved so just drop it in mesa with a one-liner patch for now.

Signed-off-by: Matthias Reichl <hias@horus.com>